Move make_zmanim() method to entity
